Simplify parsing of transation amounts
Apparantly, an Inflow/Outflow value is not at all required to have '.' as decimal separator. YNAB handles values such as "6.600,00" just fine (as long as they are inside quotes in the import csv). Copying the raw string is then a better approach since I don't  want to change the transaction value.