Parse monetary columns as Decimal
Before this commit I did not even bother to parse the monetary columns.
It was easy enough to just pass the string value directly to the
YNAB-import CSV.
However, there is a need to support multiple monetary columns per
statement row, i.e., a transaction cloumn plus a fee column.
These must be summed to calculate the net flow.
Therefore, such columns must be parsed as numerical values.
The `Decimal` module gives exact calculation without the rounding errors
of floats, which is ideal in this context.

Includes additional refactoring, such as more warnings and errors.
The `test_identity` test is actually a regression test for a bug I
discovered that the other tests did not capture since they do not look
at the output.
In any case, it's a sensible test to have.